# swagger-specs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trakt Tv",
    "version": "0.2.4"
  },
  "paths": {
...
    "/lists/{query}": {
      "get": {
        "summary": "Retrieves a list of lists",
        "parameters": [
          {
            "name": "query",
            "in": "path",
            "description": "Query parameter",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}